Remove controller_action from queue time metrics
